Change list (part 2)
This is the second change that I previously did not stage for a commit.
I first staged it for a commit with `git add .`, where `.` means
"current directory", which, in turn, means: "stage all files in the
**current directory** for the next commit. I then committed them to the
repo with: `git commit -m "Change list (part 2)" --edit` to provide this
extended message.