Exercises, small changes, exercises
To stage this commit, I executed `git add -p`
Hitting 's' (split) did not work because Git could not split this patch.
Therefore, I hit 'e' to manually **e**dit the patch and deleted the lines
that started with "+" that I did not want to be in this commit. I then
exited from the editor and committed them with:
`git commit -m "Exercises, small changes, exercises" --edit`
to add this extended commit message.